Chore: #18 TableHeader 컴포넌트 분리로 인해 폴더 생성 및 구조 변경
